Updated app.js
Updated app.js delete-order-item-form function. When deleting an order item, it will now render on the same page with the order_items listed for that order. If the order_item deleted was the last order item for that order, it will delete that order from Orders and render back to the list of orders.